restore logbackxml
