+++ spring-cloud-gcp-autoconfigure/src/test/resources/logback-test.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<configuration>
+	<property name="projectId" value="test-project"/>
+
+	<include resource="org/springframework/cloud/gcp/logging/logback-json-appender.xml"/>
+
+	<statusListener class="ch.qos.logback.core.status.NopStatusListener"/>
+
+	<appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
+		<layout class="ch.qos.logback.classic.PatternLayout">
+			<Pattern>
+				%d{yyyy-MM-dd HH:mm:ss.SSS} [%thread] %-5level %logger{36} - %msg%n
+			</Pattern>
+		</layout>
+	</appender>
+
+	<logger name="org.springframework.cloud.gcp" level="warn">
+		<appender-ref ref="STDOUT"/>
+	</logger>
+</configuration>
